pyproject.toml
[project]
name = "hil-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "langchain==0.3.26",
    "langchain-chroma>=0.1.2",
    "langchain-community>=0.3.26",
    "langchain-google-genai==2.1.10",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.1",
    "pypdf==5.9.0",
    "python-dotenv==1.0.1",
    "python-multipart>=0.0.20",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
]
